fix use sep files for in mem and redis caching
